move model related op into dedicated model op folder and try to run all local notebook
