+++ uniflow/flow/transform/transform_lmqg_flow.py
@@ -3,8 +3,8 @@ from typing import Any, Dict, Sequence
 from uniflow.constants import TRANSFORM
 from uniflow.flow.flow import Flow
 from uniflow.node.node import Node
+from uniflow.op.model.model_op import LLMDataProcessor, ModelOp
 from uniflow.op.prompt_schema import GuidedPrompt
-from uniflow.op.transform.model_op import LLMDataProcessor, ModelOp
 
 
 class TransformLMQGFlow(Flow):
